pyproject.toml
[project]
name = "aspect"
version = "0.0.1"
authors = [
  { name="Eachan Johnson", email="eachan.johnson@crick.ac.uk" },
]
description = "Deep learning with uncertainty for general purpose, chemistry, and taxonomic tasks."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = [
  "ai",
  "data",
  "deep-learning",
  "machine-learning",
]

classifiers = [  
  
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "carabiner-tools[pd,mpl]>=0.0.5.post3",
  "datasets>=3",
  "huggingface_hub",
  "numpy",
  "platformdirs>=4.4.0",
  "tqdm",
]

[project.optional-dependencies]
chem = ["chemprop>=2.1.0", "schemist>=0.0.6", "scikit-learn>=1.7"]
bio = ["vectome>=0.0.4"]
dev = ["aspect[chem,bio]", "pytest", "flake8"]

[project.urls]
"Homepage" = "https://github.com/scbirlab/aspect"
"Bug Tracker" = "https://github.com/scbirlab/aspect/issues"

[project.scripts] 
aspect = "aspect.cli_module.cli:main"

[tool.setuptools]
#packages = ["duvida"]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {aspect = ["data/*.json"]}

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[test]
deps = ["pytest>=7.4", "pytest-cov", "aspect[dev]"] 
